Fix SDHC_SEND_IF_COND CRC stop bit
CRC7 is stored on bits 1-7,
but bit 0 is actually the command stop bit.
According to the documentation, it must be set to 1.
It seems that older cards didn't really checked that,
but newer ones fail to initialize if set to 0.

This adds support for recent SDHC and SDXC cards,
typically with a capacity of 32 or 64 GB.